+++ .eslintrc.js
@@ -15,5 +15,6 @@ module.exports = {
   plugins: ['react'],
   rules: {
     'react/prop-types': 0,
+    'react/react-in-jsx-scope': 'off',
   },
 };
