+++ src/components/CardExpiration.jsx
@@ -2,7 +2,8 @@ import React, { useContext } from 'react';
 import * as S from 'styles.js';
 import ErrorMessage from 'components/ErrorMessage';
 import validator from 'validations/validator';
-import { TYPES, CardStateContext, CardDispatchContext } from 'context/CardContext';
+import { CardDispatchContext, CardStateContext } from 'store/card/CardContext';
+import { TYPES } from 'store/card/types';
 import { AutoFocusInputBox } from './AutoFocusInputBox';
 
 export default function CardExpiration({ color }) {
