+++ src/components/CardListModal.jsx
@@ -1,8 +1,9 @@
 import React, { useCallback, useContext } from 'react';
 import CardCompany from 'components/CardCompany';
 import ToastModal from 'components/ToastModal';
-import { CARD_COMPANIES } from 'constants/index';
-import { TYPES, CardStateContext, CardDispatchContext } from 'context/CardContext';
+import { CARD_COMPANIES } from 'constants/cardDomain';
+import { CardDispatchContext, CardStateContext } from 'store/card/CardContext';
+import { TYPES } from 'store/card/types';
 
 export default function CardListModal() {
   const { cardCompanyIndex, listModalFlag } = useContext(CardStateContext);
