+++ src/components/CardOwner.jsx
@@ -2,7 +2,8 @@ import React, { useContext } from 'react';
 import * as S from 'styles.js';
 import validator from 'validations/validator';
 import ErrorMessage from 'components/ErrorMessage';
-import { TYPES, CardStateContext, CardDispatchContext } from 'context/CardContext';
+import { CardDispatchContext, CardStateContext } from 'store/card/CardContext';
+import { TYPES } from 'store/card/types';
 
 export default function CardOwner({ color }) {
   const { cardOwner, cardOwnerErrorMessage } = useContext(CardStateContext);
