+++ src/components/ErrorMessage.jsx
@@ -1,6 +1,6 @@
 import React, { useContext, useEffect } from 'react';
 import * as S from 'styles.js';
-import { CardDispatchContext } from 'context/CardContext.jsx';
+import { CardDispatchContext } from 'store/card/CardContext';
 
 export default function ErrorMessage({ value, children, validate, type }) {
   const dispatch = useContext(CardDispatchContext);
