+++ src/components/ToastModal.jsx
@@ -1,7 +1,7 @@
 import React, { useCallback, useContext } from 'react';
 import * as S from 'styles.js';
 import Dimmer from 'components/Dimmer';
-import { CardDispatchContext } from 'context/CardContext';
+import { CardDispatchContext } from 'store/card/CardContext';
 
 export default function ToastModal({ type, show, children }) {
   const dispatch = useContext(CardDispatchContext);
