+++ src/store/card/reducer.js
@@ -1,51 +1,10 @@
-import React, { createContext, useReducer } from 'react';
-
-export const TYPES = {
-  SET_NUMBER: 'SET_NUMBER',
-  SET_NUMBER_ERROR_MESSAGE: 'SET_NUMBER_ERROR_MESSAGE',
-  SET_EXPIRATION: 'SET_EXPIRATION',
-  SET_EXPIRATION_ERROR_MESSAGE: 'SET_EXPIRATION_ERROR_MESSAGE',
-  SET_OWNER: 'SET_OWNER',
-  SET_OWNER_ERROR_MESSAGE: 'SET_OWNER_ERROR_MESSAGE',
-  SET_CVC: 'SET_CVC',
-  SET_CVC_ERROR_MESSAGE: 'SET_CVC_ERROR_MESSAGE',
-  SET_PASSWORD: 'SET_PASSWORD',
-  SET_PASSWORD_ERROR_MESSAGE: 'SET_PASSWORD_ERROR_MESSAGE',
-  SET_COMPANY_ERROR_MESSAGE: 'SET_COMPANY_ERROR_MESSAGE',
-  SET_LIST_MODAL_FLAG: 'SET_LIST_MODAL_FLAG',
-  SET_TIP_MODAL_FLAG: 'SET_TIP_MODAL_FLAG',
-  SET_COMPANY_INDEX: 'SET_COMPANY_INDEX',
-  SUBMIT_CARD: 'SUBMIT_CARD',
-  SET_CARD_ORDER: 'SET_CARD_ORDER',
-};
-
-export const initialState = {
-  cardNumber: ['', '', '', ''],
-  cardNumberErrorMessage: '',
-  cardExpiration: ['', ''],
-  cardExpirationErrorMessage: '',
-  cardOwner: '',
-  cardOwnerErrorMessage: '',
-  cardCvc: '',
-  cardCvcErrorMessage: '',
-  cardPassword: ['', ''],
-  cardPasswordErrorMessage: '',
-  cardCompanyIndex: -1,
-  cardNickName: '',
-  listModalFlag: false,
-  tipModalFlag: false,
-  cards: [],
-};
-
-export const CardStateContext = createContext();
-export const CardDispatchContext = createContext();
+import { TYPES } from 'store/card/types';
 
 const reducer = (state, action) => {
   switch (action.type) {
     case TYPES.SET_NUMBER: {
       const cardNumber = [...state.cardNumber];
       const { value, index } = action;
-
       cardNumber[index] = value;
 
       return {
@@ -64,7 +23,6 @@ const reducer = (state, action) => {
     case TYPES.SET_EXPIRATION: {
       const cardExpiration = [...state.cardExpiration];
       const { value, index } = action;
-
       cardExpiration[index] = value;
 
       return {
@@ -105,7 +63,6 @@ const reducer = (state, action) => {
     case TYPES.SET_PASSWORD: {
       const cardPassword = [...state.cardPassword];
       const { value, index } = action;
-
       cardPassword[index] = value;
 
       return {
@@ -174,14 +131,4 @@ const reducer = (state, action) => {
   }
 };
 
-const CardProvider = ({ children }) => {
-  const [state, dispatch] = useReducer(reducer, initialState);
-
-  return (
-    <CardStateContext.Provider value={state}>
-      <CardDispatchContext.Provider value={dispatch}>{children}</CardDispatchContext.Provider>
-    </CardStateContext.Provider>
-  );
-};
-
-export default CardProvider;
+export default reducer;
