+++ src/store/card/types.js
@@ -1,0 +1,18 @@
+export const TYPES = {
+  SET_NUMBER: 'SET_NUMBER',
+  SET_NUMBER_ERROR_MESSAGE: 'SET_NUMBER_ERROR_MESSAGE',
+  SET_EXPIRATION: 'SET_EXPIRATION',
+  SET_EXPIRATION_ERROR_MESSAGE: 'SET_EXPIRATION_ERROR_MESSAGE',
+  SET_OWNER: 'SET_OWNER',
+  SET_OWNER_ERROR_MESSAGE: 'SET_OWNER_ERROR_MESSAGE',
+  SET_CVC: 'SET_CVC',
+  SET_CVC_ERROR_MESSAGE: 'SET_CVC_ERROR_MESSAGE',
+  SET_PASSWORD: 'SET_PASSWORD',
+  SET_PASSWORD_ERROR_MESSAGE: 'SET_PASSWORD_ERROR_MESSAGE',
+  SET_COMPANY_ERROR_MESSAGE: 'SET_COMPANY_ERROR_MESSAGE',
+  SET_LIST_MODAL_FLAG: 'SET_LIST_MODAL_FLAG',
+  SET_TIP_MODAL_FLAG: 'SET_TIP_MODAL_FLAG',
+  SET_COMPANY_INDEX: 'SET_COMPANY_INDEX',
+  SUBMIT_CARD: 'SUBMIT_CARD',
+  SET_CARD_ORDER: 'SET_CARD_ORDER',
+};
